atc: avoid going through all builds of a job
Within the algorithm, if we pin the version candidates to the first
passed job that is used and then we try to match those versions to the
builds of other passed jobs. If it so happens that we pin to a version
that doesn't exist within the builds of another passed job and we don't
constrain the builds to the correct version that is missing, we might
end up going through all the builds of other passed jobs. In order to
avoid that, we are trying to go through 100 builds of each passed job at
a time, so that will avoid going through all builds of a passed job
first before moving on to the next.

We are not sure how much better this will make the scheduling time,
another approach we might try later on is to bring back the constraining
the builds of a passed job to all the inputs that are constrained to
that passed job.

[#3602]

Signed-off-by: Clara Fu <cfu@pivotal.io>